pyproject.toml

[project]
name = "mdedit"
version = "0.1.0"
description = "An editor for markdown files based on EasyMDE and PyWebView."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Camille Vuillaume", email = "camillevuillaume@gmail.com"}
]
dependencies = [
    "pywebview",
    "setuptools",
#    "PySide6",
]
[project.scripts]
mdedit = "mdedit.app:run"
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[tool.setuptools]
include-package-data = true
[tool.setuptools.packages.find]
where = ["src"]
[tool.setuptools.package-data]
"mdedit" = ["frontend/**", "frontend/*"]
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]
python_files = "test_*.py"
[tool.pylint.main]
init-hook ="""
try:
  import pylint_venv
except ImportError:
  pass
else:
  pylint_venv.inithook()
"""
[tool.pylint.format]
max-line-length = 100
[tool.black]
line-length = 100

